Send Clojure maps nested in :extra as HashMaps
Previously, if you passed something like

    {:foo {:bar "baz"}}

as `:extra` data, Sentry would receive it as

    {"foo": [["bar", "baz"]]}

There's not any real harm in this, it just looks a little
surprising/clunky in the UI.

This commit changes the handling of `:extra` maps so that nested map
values are recursively turned into `java.util.HashMap` objects. This
ensures Sentry receives those nested maps as JSON objects, which makes
the data look a bit tighter in the UI.

n.b.: It looks like the current behavior is a consequence of Clojure
maps implementing both `java.lang.Iterable` and `java.util.Map`. The
`JsonMarshaller` in raven-java[^1] treats Iterables and Maps as mutually
exclusive; it just happens to check if serializable values are Iterable
first.

[^1]: https://git.io/vyB79