Ensure all keywords are stringified by java-util-hashmappify-vals
Change java-util-hashmappify-vals to also change keywords to strings
for lone values nested under a map's direct values.

When postwalk would encounter a non-map value, the value would be
returned as-is.  This means a keyword which would be nested inside
another non-map structure would not be converted to a string.  For
maps, if the keys or values are keywords themselves, it would always
directly convert them into strings, which masked the problem.

This results in weird-looking serialization which exposes the
underlying Java Keyword object, as per #67.