Simplify java-util-hashmappify-vals function
Instead of doing the map? test twice, do it once, using postwalk.
Postwalk already traverses all objects depth-first, so there's no need
to test for map? again in the inner "f" function.

To preserve existing functionality, use walk on the outer map (or
whatever object) to prevent it from being converted to a HashMap.

Add a test for the class of the resulting value and nested values to
ensure the functionality is still as before.